add fixes

+++ src/main/resources/schema.sql
@@ -32,8 +32,8 @@ CREATE TABLE IF NOT EXISTS films
 
 CREATE TABLE IF NOT EXISTS film_genres
 (
-    film_id  INTEGER NOT NULL,
-    genre_id INTEGER NOT NULL,
+    film_id  BIGINT NOT NULL,
+    genre_id BIGINT NOT NULL,
     PRIMARY KEY (film_id, genre_id),
     FOREIGN KEY (film_id) REFERENCES films (id) ON DELETE CASCADE,
     FOREIGN KEY (genre_id) REFERENCES genres (id)
@@ -41,8 +41,8 @@ CREATE TABLE IF NOT EXISTS film_genres
 
 CREATE TABLE IF NOT EXISTS film_likes
 (
-    film_id INTEGER NOT NULL,
-    user_id INTEGER NOT NULL,
+    film_id BIGINT NOT NULL,
+    user_id BIGINT NOT NULL,
     PRIMARY KEY (film_id, user_id),
     FOREIGN KEY (film_id) REFERENCES films (id) ON DELETE CASCADE,
     FOREIGN KEY (user_id) REFERENCES users (id)
@@ -50,8 +50,8 @@ CREATE TABLE IF NOT EXISTS film_likes
 
 CREATE TABLE IF NOT EXISTS friends
 (
-    user_id   INTEGER NOT NULL,
-    friend_id INTEGER NOT NULL,
+    user_id   BIGINT NOT NULL,
+    friend_id BIGINT NOT NULL,
     PRIMARY KEY (user_id, friend_id),
     FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE,
     FOREIGN KEY (friend_id) REFERENCES users (id)
